feat: always show Quit menu item

+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/home/HomeMenuView.kt
@@ -201,10 +201,14 @@ class HomeMenuView(
                 )
             }
             HomeMenu.Item.Quit -> {
-                deleteAndQuit(
-                    activity = homeActivity,
-                    coroutineScope = homeActivity.lifecycleScope,
-                )
+                if (homeActivity.settings().shouldDeleteBrowsingDataOnQuit) {
+                    deleteAndQuit(
+                        activity = homeActivity,
+                        coroutineScope = homeActivity.lifecycleScope,
+                    )
+                } else {
+                    homeActivity.finishAndRemoveTask()
+                }
             }
             HomeMenu.Item.ReconnectSync -> {
                 navController.nav(
